style: ViewController 네이밍 변경
